+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # CHANGELOG
 
-NEXT_VERSION_BUMP: PATCH
+NEXT_VERSION_BUMP: MINOR
 ## XX.XX.XX - 20XX-XX-XX
+
+### Payments
+* [ADDED] Added optional billing address and phone fields to `CollectBankAccountConfiguration.USBankAccount`.
 
 ## 23.19.0 - 2026-09-15